feat: refatorando e movendo as actions do reducer para uma pasta separada
